requirements.txt
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
python-telegram-bot>=20.7
qrcode>=7.4.2
Pillow>=10.2.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
coverage>=7.3.0
